penarikan data
