+++ resources/views/layout/main.blade.php
@@ -89,7 +89,6 @@
 </head>
 @php
     $user = Auth::user();
-    
 @endphp
 
 <body class="hold-transition sidebar-mini layout-fixed">
@@ -267,349 +266,7 @@
     <script src="{{ asset('lte/plugins/datatables-buttons/js/buttons.html5.min.js') }}"></script>
     <script src="{{ asset('lte/plugins/datatables-buttons/js/buttons.print.min.js') }}"></script>
     <script src="{{ asset('lte/plugins/datatables-buttons/js/buttons.colVis.min.js') }}"></script>
-    <script>
-        $(document).ready(function() {
-            $('#karyawanTable').DataTable({
-                "paging": true,
-                "lengthChange": true,
-                "searching": true,
-                "ordering": true,
-                "info": true,
-                "autoWidth": false,
-                "responsive": true,
-            });
-        });
-    </script>
-    <script>
-        $(document).ready(function() {
-            $('#allTable').DataTable({
-                "info": true,
-                "autoWidth": false,
-                "responsive": true,
-            });
-        });
-    </script>
-    <script>
-        $('#example2').DataTable({
-            "paging": true,
-            "lengthChange": false,
-            "searching": true,
-            "ordering": true,
-            "info": true,
-            "autoWidth": false,
-            "responsive": true,
-        });
-    </script>
-    <!-- Page Select script -->
-
-        <!-- Cuti -->
-        <script>
-            // Mendapatkan tanggal sekarang
-            var currentDate = new Date();
-            // Menambahkan 5 hari ke tanggal sekarang
-            var targetDate = new Date(currentDate);
-            targetDate.setDate(currentDate.getDate() + 5);
-        
-            // Inisialisasi datetimepicker untuk elemen input dengan id "start_date5"
-            $('#start_date5').datetimepicker({
-                format: 'YYYY-MM-DD', // Format tanggal yang diinginkan
-                icons: {
-                    time: 'fa fa-clock',
-                    date: 'fa fa-calendar',
-                    up: 'fa fa-chevron-up',
-                    down: 'fa fa-chevron-down',
-                    previous: 'fa fa-chevron-left',
-                    next: 'fa fa-chevron-right',
-                    today: 'fa fa-calendar-check-o',
-                    clear: 'fa fa-trash',
-                    close: 'fa fa-times'
-                },
-                minDate: targetDate, // Tidak memungkinkan pemilihan tanggal sebelum tanggal sekarang
-                maxDate: targetDate // Tidak memungkinkan pemilihan tanggal lebih dari 5 hari ke depan
-            });
-        
-            // Inisialisasi datetimepicker untuk elemen input dengan id "end_date5"
-            $('#end_date5').datetimepicker({
-                format: 'YYYY-MM-DD', // Format tanggal yang diinginkan
-                icons: {
-                    time: 'fa fa-clock',
-                    date: 'fa fa-calendar',
-                    up: 'fa fa-chevron-up',
-                    down: 'fa fa-chevron-down',
-                    previous: 'fa fa-chevron-left',
-                    next: 'fa fa-chevron-right',
-                    today: 'fa fa-calendar-check-o',
-                    clear: 'fa fa-trash',
-                    close: 'fa fa-times'
-                },
-                useCurrent: false, // Tidak menggunakan tanggal saat ini secara default
-                minDate: targetDate // Mengatur tanggal minimum menjadi targetDate
-            });
-        
-            // Mengatur bahwa tanggal di end_date5 tidak bisa sebelum tanggal di start_date5
-            $("#start_date5").on("change.datetimepicker", function (e) {
-                $('#end_date5').datetimepicker('minDate', e.date);
-            });
-        
-            // Mengatur bahwa tanggal di start_date5 tidak bisa setelah tanggal di end_date5
-            $("#end_date5").on("change.datetimepicker", function (e) {
-                $('#start_date5').datetimepicker('maxDate', e.date);
-            });
-        </script>
-
-        <!-- Overtime -->
-        <script>
-            // Inisialisasi datetimepicker untuk elemen input dengan id "start_dateover"
-            $('#start_dateover').datetimepicker({
-                format: 'YYYY-MM-DD HH:mm', // Format tanggal dan waktu yang diinginkan
-                icons: {
-                    time: 'fa fa-clock',
-                    date: 'fa fa-calendar',
-                    up: 'fa fa-chevron-up',
-                    down: 'fa fa-chevron-down',
-                    previous: 'fa fa-chevron-left',
-                    next: 'fa fa-chevron-right',
-                    today: 'fa fa-calendar-check-o',
-                    clear: 'fa fa-trash',
-                    close: 'fa fa-check'
-                },
-                sideBySide: false, // Menampilkan input waktu secara berdampingan dengan input tanggal
-                toolbarPlacement: 'bottom', // Menempatkan toolbar di bagian bawah
-                buttons: {
-                    showClose: true, // Menampilkan tombol Close
-                    showToday: true, // Menampilkan tombol Today
-                    showClear: true, // Menampilkan tombol Clear
-                    showApply: true // Menampilkan tombol Apply
-                }
-            });
-        
-            // Inisialisasi datetimepicker untuk elemen input dengan id "end_dateover"
-            $('#end_dateover').datetimepicker({
-                format: 'YYYY-MM-DD HH:mm', // Format tanggal dan waktu yang diinginkan
-                icons: {
-                    time: 'fa fa-clock',
-                    date: 'fa fa-calendar',
-                    up: 'fa fa-chevron-up',
-                    down: 'fa fa-chevron-down',
-                    previous: 'fa fa-chevron-left',
-                    next: 'fa fa-chevron-right',
-                    today: 'fa fa-calendar-check-o',
-                    clear: 'fa fa-trash',
-                    close: 'fa fa-check'
-                },
-                useCurrent: false, // Tidak menggunakan tanggal saat ini secara default
-                sideBySide: false, // Menampilkan input waktu secara berdampingan dengan input tanggal
-                toolbarPlacement: 'bottom', // Menempatkan toolbar di bagian bawah
-                buttons: {
-                    showClose: true, // Menampilkan tombol Close
-                    showToday: true, // Menampilkan tombol Today
-                    showClear: true, // Menampilkan tombol Clear
-                    showApply: true // Menampilkan tombol Apply
-                }
-            });
-        
-            // Mengatur bahwa tanggal di end_dateover tidak bisa sebelum tanggal di start_dateover
-            $("#start_dateover").on("change.datetimepicker", function (e) {
-                $('#end_dateover').datetimepicker('minDate', e.date);
-            });
-        
-            // Mengatur bahwa tanggal di start_dateover tidak bisa setelah tanggal di end_dateover
-            $("#end_dateover").on("change.datetimepicker", function (e) {
-                $('#start_dateover').datetimepicker('maxDate', e.date);
-            });
-        </script>
-        
-        
-    
-        <!-- Cuti -->
-            <script>
-                // Mendapatkan tanggal sekarang
-                var currentDate = new Date();
-
-                // Inisialisasi datetimepicker untuk elemen input dengan id "start_dateabsen"
-                $('#start_date').datetimepicker({
-                    format: 'YYYY-MM-DD', // Format tanggal yang diinginkan
-                    icons: {
-                        time: 'fa fa-clock',
-                        date: 'fa fa-calendar',
-                        up: 'fa fa-chevron-up',
-                        down: 'fa fa-chevron-down',
-                        previous: 'fa fa-chevron-left',
-                        next: 'fa fa-chevron-right',
-                        today: 'fa fa-calendar-check-o',
-                        clear: 'fa fa-trash',
-                        close: 'fa fa-times'
-                    },
-                
-                });
-
-                $('#time_in').datetimepicker({
-                    format: 'HH:mm'
-                });
-                // Inisialisasi datetimepicker untuk elemen input dengan id "end_date"
-                $('#end_date').datetimepicker({
-                    format: 'YYYY-MM-DD', // Format tanggal yang diinginkan
-                    icons: {
-                        time: 'fa fa-clock',
-                        date: 'fa fa-calendar',
-                        up: 'fa fa-chevron-up',
-                        down: 'fa fa-chevron-down',
-                        previous: 'fa fa-chevron-left',
-                        next: 'fa fa-chevron-right',
-                        today: 'fa fa-calendar-check-o',
-                        clear: 'fa fa-trash',
-                        close: 'fa fa-times'
-                    },
-                    useCurrent: false // Tidak menggunakan tanggal saat ini secara default
-                });
-
-                // Mengatur bahwa tanggal di end_date tidak bisa sebelum tanggal di start_date
-                $("#start_date").on("change.datetimepicker", function (e) {
-                    $('#end_date').datetimepicker('minDate', e.date);
-                });
-
-                // Mengatur bahwa tanggal di start_date tidak bisa setelah tanggal di end_date
-                $("#end_date").on("change.datetimepicker", function (e) {
-                    $('#start_date').datetimepicker('maxDate', e.date);
-                });
-            </script>
-        <!-- //Cuti -->
-        <!-- Absen -->
-            <script>
-                // Mendapatkan tanggal sekarang
-                var currentDate = new Date();
-
-                // Inisialisasi datetimepicker untuk elemen input dengan id "start_dateabsen"
-                $('#start_dateabsen').datetimepicker({
-                    format: 'YYYY-MM-DD', // Format tanggal yang diinginkan
-                    icons: {
-                        time: 'fa fa-clock',
-                        date: 'fa fa-calendar',
-                        up: 'fa fa-chevron-up',
-                        down: 'fa fa-chevron-down',
-                        previous: 'fa fa-chevron-left',
-                        next: 'fa fa-chevron-right',
-                        today: 'fa fa-calendar-check-o',
-                        clear: 'fa fa-trash',
-                        close: 'fa fa-times'
-                    },
-                
-                });
-
-                // Inisialisasi datetimepicker untuk elemen input dengan id "end_dateabsen"
-                $('#end_dateabsen').datetimepicker({
-                    format: 'YYYY-MM-DD', // Format tanggal yang diinginkan
-                    icons: {
-                        time: 'fa fa-clock',
-                        date: 'fa fa-calendar',
-                        up: 'fa fa-chevron-up',
-                        down: 'fa fa-chevron-down',
-                        previous: 'fa fa-chevron-left',
-                        next: 'fa fa-chevron-right',
-                        today: 'fa fa-calendar-check-o',
-                        clear: 'fa fa-trash',
-                        close: 'fa fa-times'
-                    },
-                    useCurrent: false // Tidak menggunakan tanggal saat ini secara default
-                });
-
-                // Mengatur bahwa tanggal di end_dateabsen tidak bisa sebelum tanggal di start_dateabsen
-                $("#start_dateabsen").on("change.datetimepicker", function (e) {
-                    $('#end_dateabsen').datetimepicker('minDate', e.date);
-                });
-
-                // Mengatur bahwa tanggal di start_dateabsen tidak bisa setelah tanggal di end_dateabsen
-                $("#end_dateabsen").on("change.datetimepicker", function (e) {
-                    $('#start_dateabsen').datetimepicker('maxDate', e.date);
-                });
-            </script>
-        <!-- Absen -->
-    
-    <script>
-        $(function () {
-          //Initialize Select2 Elements
-          $('.select2').select2()
-      
-          //Initialize Select2 Elements
-          $('.select2bs4').select2({
-            theme: 'bootstrap4'
-          })
-      
-          //Datemask dd/mm/yyyy
-          $('#datemask').inputmask('dd/mm/yyyy', { 'placeholder': 'dd/mm/yyyy' })
-          //Datemask2 mm/dd/yyyy
-          $('#datemask2').inputmask('mm/dd/yyyy', { 'placeholder': 'mm/dd/yyyy' })
-          //Money Euro
-          $('[data-mask]').inputmask()
-      
-          //Date picker
-          $('#reservationdate').datetimepicker({
-              format: 'L',
-              locale: 'id'
-          });
-      
-          //Date and time picker
-          $('#reservationdatetime').datetimepicker({ icons: { time: 'far fa-clock' } });
-      
-          //Date range picker
-          $('#reservation').daterangepicker()
-          //Date range picker with time picker
-          $('#reservationtime').daterangepicker({
-            timePicker: true,
-            timePickerIncrement: 30,
-            locale: {
-              format: 'MM/DD/YYYY hh:mm A'
-            }
-          })
-          //Date range as a button
-          $('#daterange-btn').daterangepicker(
-            {
-              ranges   : {
-                'Today'       : [moment(), moment()],
-                'Yesterday'   : [moment().subtract(1, 'days'), moment().subtract(1, 'days')],
-                'Last 7 Days' : [moment().subtract(6, 'days'), moment()],
-                'Last 30 Days': [moment().subtract(29, 'days'), moment()],
-                'This Month'  : [moment().startOf('month'), moment().endOf('month')],
-                'Last Month'  : [moment().subtract(1, 'month').startOf('month'), moment().subtract(1, 'month').endOf('month')]
-              },
-              startDate: moment().subtract(29, 'days'),
-              endDate  : moment()
-            },
-            function (start, end) {
-              $('#reportrange span').html(start.format('MMMM D, YYYY') + ' - ' + end.format('MMMM D, YYYY'))
-            }
-          )
-      
-          //Timepicker
-          $('#timepicker').datetimepicker({
-            format: 'LT'
-          })
-
-          $('#time_in').datetimepicker({
-            format: 'HH:mm'
-          })
-      
-          //Bootstrap Duallistbox
-          $('.duallistbox').bootstrapDualListbox()
-      
-          //Colorpicker
-          $('.my-colorpicker1').colorpicker()
-          //color picker with addon
-          $('.my-colorpicker2').colorpicker()
-      
-          $('.my-colorpicker2').on('colorpickerChange', function(event) {
-            $('.my-colorpicker2 .fa-square').css('color', event.color.toString());
-          })
-      
-          $("input[data-bootstrap-switch]").each(function(){
-            $(this).bootstrapSwitch('state', $(this).prop('checked'));
-          })
-      
-        })
-        
-      
-      
-      </script>
+    <script src="{{ asset('js/rs.js') }}"></script>
     <!-- SweetAlert2 -->
     <script src="{{ asset('lte/plugins/sweetalert2/sweetalert2.min.js') }}"></script>
     <!-- Page specific script SweetAlert2-->
@@ -633,329 +290,32 @@
             });
         </script>
     @endif
-    
+
+    @role('karyawan')
     <script>
-        $(function () {
-            var today = new Date().toISOString().slice(0, 10); // Mendapatkan tanggal hari ini dalam format YYYY-MM-DD
-    
-            $("#laporan").DataTable({
-                "responsive": true,
-                "lengthChange": false,
-                "autoWidth": false,
-                "buttons": [
-                    {
-                        extend: 'excel',
-                        filename: function() {
-                            return 'Laporan_Cuti_' + today; // Menetapkan nama file sebagai "Laporan_Cuti_tanggal_hari_ini"
+        document.addEventListener('DOMContentLoaded', function() {
+            function updateBadge(url, badgeId, dataKey) {
+                fetch(url)
+                    .then(response => response.json())
+                    .then(data => {
+                        const badge = document.getElementById(badgeId);
+                        if (data[dataKey] > 0) {
+                            badge.textContent = data[dataKey];
+                            badge.style.display = 'inline-block'; // Menampilkan badge jika nilainya lebih dari 0
+                        } else {
+                            badge.style.display = 'none'; // Menyembunyikan badge jika nilainya 0
                         }
-                    }
-                ]
-            }).buttons().container().appendTo('#laporan_wrapper .col-md-6:eq(0)');
-    
-            $("#laporan_lembur").DataTable({
-                "responsive": true,
-                "lengthChange": false,
-                "autoWidth": false,
-                "buttons": [
-                    {
-                        extend: 'excel',
-                        filename: function() {
-                            return 'Laporan_Lembur_' + today; // Menetapkan nama file sebagai "Laporan_Lembur_tanggal_hari_ini"
-                        }
-                    }
-                ]
-            }).buttons().container().appendTo('#laporan_lembur_wrapper .col-md-6:eq(0)');
-            
-            $("#laporan_oncall").DataTable({
-                "responsive": true,
-                "lengthChange": false,
-                "autoWidth": false,
-                "buttons": [
-                    {
-                        extend: 'excel',
-                        filename: function() {
-                            return 'Laporan_Oncall_' + today; // Menetapkan nama file sebagai "Laporan_Lembur_tanggal_hari_ini"
-                        }
-                    }
-                ]
-            }).buttons().container().appendTo('#laporan_oncall_wrapper .col-md-6:eq(0)');
-
-            $("#lap_absensi").DataTable({
-            "responsive": true,
-            "lengthChange": false,
-            "autoWidth": false,
-            "buttons": [
-                {
-                    extend: 'excelHtml5',
-                    text: 'Export to Excel',
-                    filename: 'Laporan_Absensi_' + today, // Nama file ekspor
-                }
-            ]
-        }).buttons().container().appendTo('#lap_absensi_wrapper .col-md-6:eq(0)');
-        });
-    </script>
-<!-- TAMBAH/EDIT cuti -->
-<script>
-$(document).ready(function() {
-    // Handle kategori_cuti change event
-    $('#kategori_cuti').change(function() {
-        var kategoriCuti = $(this).val();
-        if (kategoriCuti) {
-            $.ajax({
-                url: '/pengajuan-cuti/create/' + kategoriCuti,
-                type: 'GET',
-                success: function(data) {
-                    $('#leave_type_id').empty();
-                    $('#leave_type_id').append('<option value="" disabled selected>Pilih Jenis Cuti</option>');
-                    $.each(data, function(key, value) {
-                        $('#leave_type_id').append('<option value="' + key + '">' + value + '</option>');
-                    });
-                    $('#leave_type_id_container').show();
-
-                    // Show/hide file upload container based on kategori_cuti
-                    if (kategoriCuti === 'CUTI TAHUNAN') {
-                        $('#leave_type_id_container').hide();
-                        $('#leave_type_id').val('20'); // Set leave_type_id value to 20
-                        $('#max_amount_display').text('Maksimal Jumlah Cuti: 5').show();
-                    } else {
-                        $('#leave_type_id_container').show();
-                        $('#max_amount_display').hide();
-                    }
-
-                    if (kategoriCuti === 'CUTI KHUSUS') {
-                        $('#file_upload_container').show();
-                        $('#file_upload').prop('required', true);
-                    } else {
-                        $('#file_upload_container').hide();
-                        $('#file_upload').prop('required', false);
-                    }
-                }
-            });
-        } else {
-            $('#leave_type_id_container').hide();
-            $('#file_upload_container').hide();
-            $('#file_upload').prop('required', false);
-            $('#max_amount_display').hide();
-        }
-    });
-
-    // Handle leave_type_id change event
-    $('#leave_type_id').change(function() {
-        var leaveTypeId = $(this).val();
-        if (leaveTypeId) {
-            $.ajax({
-                url: '/pengajuan-cuti/leave-types/' + leaveTypeId,
-                type: 'GET',
-                success: function(data) {
-                    if (data.max_amount) {
-                        $('#max_amount_display').text('Maksimal Jumlah Cuti: ' + data.max_amount + ' Hari').show();
-                    } else {
-                        $('#max_amount_display').hide();
-                    }
-
-                    if (data.file_upload === 'yes') {
-                        $('#file_upload_container').show();
-                        $('#file_upload').prop('required', true);
-                    } else {
-                        $('#file_upload_container').hide();
-                        $('#file_upload').prop('required', false);
-                    }
-                }
-            });
-        } else {
-            $('#max_amount_display').hide();
-            $('#file_upload_container').hide();
-            $('#file_upload').prop('required', false);
-        }
-    });
-
-
-});
-
-    // Handle file upload change event to update the label with the selected file name
-    $('#file_upload').change(function() {
-        var fileName = $(this).val().split('\\').pop(); // Extract the file name from the file path
-        $(this).next('.custom-file-label').html(fileName); // Update the label text
-    });
-
-$(document).ready(function() {
-    $('.select2bst4').select2({
-        theme: 'bootstrap4'
-    });
-});
-</script> 
- 
-
-<!-- Menampilkan kalkulasi waktu -->
-<script type="text/javascript">
-    $(document).ready(function() {
-        // Initialize datetime pickers
-        $('#start_dateover').datetimepicker({
-            format: 'YYYY-MM-DD HH:mm'
-        });
-        $('#end_dateover').datetimepicker({
-            format: 'YYYY-MM-DD HH:mm',
-            useCurrent: false
-        });
-
-        // Calculate the difference in hours and minutes when either date changes
-        $('#start_dateover, #end_dateover').on('change.datetimepicker', function() {
-            var startDate = $('#start_dateover').find('input').val();
-            var endDate = $('#end_dateover').find('input').val();
-
-            if (startDate && endDate) {
-                var start = moment(startDate, 'YYYY-MM-DD HH:mm');
-                var end = moment(endDate, 'YYYY-MM-DD HH:mm');
-                
-                // Calculate the difference in minutes
-                var diff = end.diff(start, 'minutes');
-
-                // Convert total minutes to hours and minutes
-                var hours = Math.floor(diff / 60);
-                var minutes = diff % 60;
-
-                $('#total_duration').val(hours + ' jam ' + minutes + ' menit');
+                    })
+                    .catch(() => {}); 
             }
-        });
-    });
-</script>
 
-<script>
-    $(function () {
-        // Initialize the date pickers
-        $('#start_date').datetimepicker({
-            format: 'L'
-        });
-        $('#end_date').datetimepicker({
-            format: 'L'
-        });
-
-        // Update total days when either date changes
-        $('#start_date, #end_date').on('change.datetimepicker', function () {
-            calculateDays();
-        });
-
-        function calculateDays() {
-            var startDate = $('#start_date').datetimepicker('date');
-            var endDate = $('#end_date').datetimepicker('date');
-            
-            if (startDate && endDate) {
-                var start = moment(startDate);
-                var end = moment(endDate);
-                var days = end.diff(start, 'days') + 1; // Add 1 to include both start and end dates
-
-                $('#total_days').val(days + ' Hari');
-            }
-        }
-    });
-</script>
-
-@role('Super-Admin|karyawan')
-
-<script>
-    document.addEventListener('DOMContentLoaded', function() {
-        function updateBadge(url, badgeId, dataKey) {
-            fetch(url)
-                .then(response => response.json())
-                .then(data => {
-                    const badge = document.getElementById(badgeId);
-                    if (data[dataKey] > 0) {
-                        badge.textContent = data[dataKey];
-                        badge.style.display = 'inline-block'; // Menampilkan badge jika nilainya lebih dari 0
-                    } else {
-                        badge.style.display = 'none'; // Menyembunyikan badge jika nilainya 0
-                    }
-                })
-                .catch(() => {}); 
-        }
-
-        // Initial calls to update all counts
-        updateBadge('{{ route('api.pending-count') }}', 'pendingCountBadge', 'pendingCount');
-        updateBadge('{{ route('api.over-count') }}', 'lemburCountBadge', 'countOvertime');
-        updateBadge('{{ route('api.oncall-count') }}', 'oncallCountBadge', 'countOncall');
-
-        // Optionally, update the counts periodically
-        setInterval(function() {
+            // Initial calls to update all counts
             updateBadge('{{ route('api.pending-count') }}', 'pendingCountBadge', 'pendingCount');
             updateBadge('{{ route('api.over-count') }}', 'lemburCountBadge', 'countOvertime');
             updateBadge('{{ route('api.oncall-count') }}', 'oncallCountBadge', 'countOncall');
-        }, 5000); // Update every minute
-    });
-</script>
-@endrole
-<script>
-    // Show input fields when pelatihan is selected
-    $('#pelatihan').on('change', function() {
-        var selected = $(this).val();
-        $('.pelatihan-group').hide(); // Hide all
-        selected.forEach(function(id) {
-            if ($('#pelatihan-' + id + '-details').length) {
-                // Show existing pelatihan details
-                $('#pelatihan-' + id + '-details').show();
-            } else {
-                // Add new pelatihan details if not already present
-                addPelatihanDetails(id);
-            }
         });
-    });
-
-    // Initially show fields for already selected pelatihan
-    $(document).ready(function() {
-        var selected = $('#pelatihan').val();
-        selected.forEach(function(id) {
-            if ($('#pelatihan-' + id + '-details').length) {
-                $('#pelatihan-' + id + '-details').show(); // Show existing selected on load
-            }
-        });
-    });
-
-    // Function to add new pelatihan details dynamically
-    function addPelatihanDetails(id) {
-        var pelatihanName = $('#pelatihan option[value="' + id + '"]').text().trim(); // Get selected pelatihan name
-        var pelatihanDetails = `
-            <div class="form-group pelatihan-group" id="pelatihan-` + id + `-details">
-                <!-- Edit Nama Pelatihan -->
-                <label for="name_` + id + `">Nama Pelatihan untuk ` + pelatihanName + `</label>
-                <input type="text" name="nama_pelatihan[` + id + `]" value="` + pelatihanName + `" class="form-control" disabled>
-                
-                <!-- Tanggal Expired Pelatihan -->
-                <label for="tanggal_expired_` + id + `">Tanggal Expired untuk ` + pelatihanName + `</label>
-                <input type="date" name="tanggal_expired[` + id + `]" class="form-control">
-                
-                <!-- File Upload Pelatihan -->
-                <label for="file_` + id + `">File Sertifikat untuk ` + pelatihanName + `</label>
-                <input type="file" name="file[` + id + `]" class="form-control" accept=".pdf">
-                <hr style="border: 2px solid #000;">
-            </div>
-        `;
-        $('#pelatihan-details').append(pelatihanDetails);
-    }
-
-    // Add new Pelatihan input field
-    $('#add-pelatihan').on('click', function() {
-        // Set the flag to true when new pelatihan is added
-        $('#add_pelatihan_flag').val('true');
-
-        var newPelatihanFields = `
-            <div class="form-group new-pelatihan-div">
-                <input type="text" name="new_pelatihan[]" class="form-control mb-2 new-pelatihan" placeholder="Nama Pelatihan Baru" required>
-                <label for="new_tanggal_expired[]">Tanggal Expired Pelatihan Baru</label>
-                <input type="date" name="new_tanggal_expired[]" class="form-control mb-2 new-expired">
-                <label for="new_file[]">File Sertifikat Pelatihan Baru</label>
-                <input type="file" name="new_file[]" class="form-control mb-2 new-file" accept=".pdf">
-                <!-- Cancel button to remove the new pelatihan div -->
-                <button type="button" class="btn btn-danger remove-pelatihan-btn">Batal</button>
-                <hr style="border: 2px solid #000;">
-            </div>
-        `;
-        $('#new-pelatihan-container').append(newPelatihanFields);
-    });
-
-    // Function to remove the new pelatihan fieldset
-    $(document).on('click', '.remove-pelatihan-btn', function() {
-        $(this).closest('.new-pelatihan-div').remove();
-    });
-</script>
+    </script>
+    @endrole
 
 </body>
