+++ resources/views/oncall/report-history.blade.php
@@ -1,0 +1,68 @@
+@extends('layout.main')
+@section('content')
+<div class="content-wrapper">
+    <!-- Content Header (Page header) -->
+    <div class="content-header">
+        <div class="container-fluid">
+            <div class="row mb-2">
+                <div class="col-sm-6">
+                    <h1 class="m-0">OnCall</h1>
+                </div><!-- /.col -->
+                <div class="col-sm-6">
+                    <ol class="breadcrumb float-sm-right">
+                        <li class="breadcrumb-item"><a href="#">Riwayat Penarikan OnCall</a></li>
+                        <li class="breadcrumb-item active">Daftar Hadir</li>
+                    </ol>
+                </div><!-- /.col -->
+            </div><!-- /.row -->
+        </div><!-- /.container-fluid -->
+    </div>
+    <!-- /.content-header -->
+
+    <!-- Main content -->
+    <section class="content">
+        <div class="container-fluid">
+            <div class="row">
+                <div class="col-12">
+                    <div class="card">
+                        <div class="card-header">
+                        </div>
+                        <!-- /.card-header -->
+                        <div class="card-body ">
+                            <table class="table table-bordered table-hover" id="allTable">
+                                <thead>
+                                    <tr>
+                                        <th style="width: 10px">No</th>
+                                        <th>Nama User</th>
+                                        <th>Waktu Penarikan</th>
+                                        <th>Tanggal Awal</th>
+                                        <th>Tanggal Akhir</th>
+                                        <th>Ip Address</th>
+                                    </tr>
+                                </thead>
+                                <tbody>
+                                    @foreach ($reporthistory as $oncall)
+                                    <tr>
+                                        <td>{{ $loop->iteration }}</td>
+                                        <td>{{ $oncall->user->name }}</td>
+                                        <td>{{ $oncall->created_at }}</td>
+                                        <td>{{ $oncall->start_date }}</td>
+                                        <td>{{ $oncall->end_date }}</td>
+                                        <td>{{ $oncall->ip_address }}</td>
+                                    </tr>
+                                    @endforeach
+                                </tbody>
+                            </table>
+                        </div>
+                        <!-- /.card-body -->
+                    </div>
+                    <!-- /.card -->
+                </div>
+            </div>
+            <!-- /.row (main row) -->
+        </div><!-- /.container-fluid -->
+    </section>
+    <!-- /.content -->
+</div>
+
+@endsection
